+++ stdext/list.mli
@@ -14,3 +14,4 @@ val max_exn : compare:('a -> 'a -> int) -> 'a list -> 'a
     Raises Invalig_argument on empty lists *)
 
 val compare : compare:('a -> 'a -> int) -> 'a list -> 'a list -> int
+val unzip : ('a * 'b) list -> 'a list * 'b list